move new report to reporting module
