+++ pyproject.toml
@@ -49,7 +49,7 @@ lt_status = 'mozilla_bitbar_devicepool.lambdatest.status:lt_status_main'
 lt_compare_to_api = 'mozilla_bitbar_devicepool.compare_to_api_lt:main'
 lt_failure_report = 'mozilla_bitbar_devicepool.lambdatest.status:lt_failed_job_report'
 lt_success_rate_report = 'mozilla_bitbar_devicepool.lambdatest.status:lt_success_rate_report'
-lt_job_distribution_report = 'mozilla_bitbar_devicepool.lambdatest.status:job_distribution_report'
+lt_job_distribution_report = 'mozilla_bitbar_devicepool.reporting.main:job_distribution_report'
 
 [build-system]
 requires = ["poetry-core"]
